Added AdvancedRemoteListener interface.
The existing callback system uses function pointers; this doesn't work for class methods.  To make it easier/possible for a object to be notified of events, this commit adds an AdvancedRemoteListener interface.  Objects wishing to be notified of events need to implement this interface.  The listener takes precedence over function pointers.